GS4430Y: DTS changes for enabling WiLink8.
Update the regulator node for wlan and add a regulator for bt to
enable WiLink8 on Duovero. Enable pin muxing for bt which is
connected over UART2.

The default mmc clock speed of 50MHz results in kernel panic and
sdio driver is unable to talk to the wlcore driver. Changing the
max-frequency of mmc bus to 40MHz fixes this problem.

Signed-off-by: Arun Bharadwaj <arun@gumstix.com>
Signed-off-by: Ash Charles <ashcharles@gmail.com>