edt-ft5x06: Handle reset/wake for different versions of Newhaven displays
On older versions of some Newhaven displays using the edt-ft5x06
capative touch controller, the 'WAKE' pin is exposed on the flex
connector. On newer versions of the same displays, the 'RESET' pin
is broken out at the same place instead. As the screen needs to be
reset before any version information can be read, treat a 'wake-gpio'
from the device tree as the reset pin (if one is not otherwise defined)
to accommodate new and old screen versions with the same device tree.

Signed-off-by: Jerry Hung <jerry@gumstix.com>